requirements.txt
django
djangorestframework
Pillow
psycopg2-binary
decouple
